new: support for injecting open tracing tracker in publication [major]
This patch allows to send tracking information in bahamut.Publication.

In order to do so, it wires the open tracing span in the
bahamut.Publication object so it can be decoded after going through the
pubsub engine.

**BREAKING CHANGE** now the entire publication is sent to the PubSub,
and not only the publication data. This is the only way to carry the
needed additional information. All service publishing/subscribing using
a bahamut.PubSubServer must be using at least this version.