[major] new: Bahamut now takes a list of Authenticators and Authorizers
This patch modifies the bahamut config in order to be able to pass a list of Authenticators/Authorizers. They are executed in order, and for the request to be authenticated/authorized, all the ones in the list must return true. If one returns false or an error, the request will be rejected.