new: adds ability to distinguish between different Publish requests when using the NATS request-reply messaging
• NATSOptPublishRequireAck option 
When configured, it will set the ResponseMode field of the Publication to “ReplyWithACK”. The side effect of this is that caller of Publish will now block until an ACK is received or the supplied context deadline is reached. As a result of this option our implementation of Subscribe will now need to look for this new ResponseMode attribute when consuming messages to determine whether it needs to respond back with an ACK - the value would be set to “ReplyWithACK”. This option is mutually exclusive with the NATSOptRespondToChannel option.

• NATSOptRespondToChannel option 
When configured, it will set the ResponseMode field of the Publication to “ReplyWithPublication”. The side effect of this is that caller of Publish will now block until a Publication response is received or the supplied context deadline is reached. As a result of this option our implementation of Subscribe will now need to look for the new ResponseMode attribute and check to see if its value is set to “ReplyWithPublication” as this would imply that a Publication response should be delivered AFTER processing has been completed. This option is mutually exclusive with the NATSOptPublishRequireAck option.

breaking: removed the NATSOptPublishReplyValidator publish option as it now serves no purpose. Folks should either use NATSOptPublishRequireAck or NATSOptRespondToChannel now. As a result, the replyValidator attribute from the publish configuration has been removed as well.
Signed-off-by: Amir Aslaminejad <amir.a@aporeto.com>