fixed/metrics/prometheus: correctly handle prefix (#178)
This patches fixes a bug where the sanitization function of the
prometheus metric manager was not correctly handling the given path if
it was containing a '_prefix'. This patch fixes this issue and adds
tests.

The parameter 'url' of the interface MetricManager.MeasureRequest has
also been renamed to 'path' for better usage clarity.